[DRAFT] Spring integration

+++ src/main/java/am/aca/generactive/service/ItemServiceImpl.java
@@ -1,60 +1,71 @@
 package am.aca.generactive.service;
 
 import am.aca.generactive.model.Item;
-import am.aca.generactive.repository.BasketRepository;
 import am.aca.generactive.repository.ItemRepository;
-import org.hibernate.Session;
-import org.hibernate.SessionFactory;
-import org.hibernate.Transaction;
-import org.springframework.stereotype.Component;
+import am.aca.generactive.service.dto.ItemDTO;
+import am.aca.generactive.service.mapper.ItemDTOMapper;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.data.jpa.domain.Specification;
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
 import java.util.Optional;
 
-@Component
+@Service
 public class ItemServiceImpl implements ItemService {
 
     private final ItemRepository itemRepository;
-    private final BasketRepository basketRepository;
-    private final SessionFactory sessionFactory;
 
-    public ItemServiceImpl(ItemRepository itemRepository,
-                           BasketRepository basketRepository, SessionFactory sessionFactory) {
+    @Autowired
+    public ItemServiceImpl(ItemRepository itemRepository) {
         this.itemRepository = itemRepository;
-        this.basketRepository = basketRepository;
-        this.sessionFactory = sessionFactory;
     }
 
     @Override
-    public Item create(Item item) {
-        Session session = sessionFactory.getCurrentSession();
-        Transaction transaction = session.beginTransaction();
+    @Transactional
+    public ItemDTO create(ItemDTO item) {
+        Item entity = ItemDTOMapper.mapToEntity(item);
 
-        itemRepository.insert(item);
-        basketRepository.getBasket(null);
+        itemRepository.save(entity);
 
-        transaction.commit();
-        session.close();
-        return item;
+        return ItemDTOMapper.mapToDTO(entity).orElse(null);
     }
 
     @Override
     public Item update(Item item) {
-        return itemRepository.update(item);
+        return itemRepository.save(item);
     }
 
     @Override
     public boolean delete(Long id) {
-        return itemRepository.deleteById(id);
+        if (!itemRepository.existsById(id)) {
+            return false;
+        }
+
+        itemRepository.deleteById(id);
+
+        return true;
     }
 
     @Override
-    public Optional<Item> getItem(Long id) {
-        return itemRepository.findById(id);
+    public Optional<ItemDTO> getItem(Long id) {
+        Optional<Item> item = itemRepository.findById(id);
+
+        return ItemDTOMapper.mapToDTO(item.orElse(null));
     }
 
     @Override
-    public List<? extends Item> getAll() {
-        return itemRepository.getAllItems();
+    public List<? extends ItemDTO> getAll() {
+        return ItemDTOMapper.mapToDTOs(itemRepository.findAll());
+    }
+
+    @Override
+    public List<? extends ItemDTO> find(String name) {
+        Specification<Item> specification = Specification.where(null);
+
+        return ItemDTOMapper
+                .mapToDTOs(itemRepository
+                        .find(name, 300));
     }
 }
